+++ legateboost/special.py
@@ -18,8 +18,9 @@ from enum import IntEnum
 
 import cunumeric as cn
 from legate.core import get_legate_runtime, types as ty
-from legateboost.library import user_context, user_lib
-from legateboost.utils import get_store
+
+from .library import user_context, user_lib
+from .utils import get_store
 
 
 class _SpecialOpCode(IntEnum):
